added CPP_02

+++ CPP_00/ex01/Contact.hpp
@@ -6,7 +6,7 @@
 /*   By: aoo <aoo@student.42singapore.sg>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/21 03:06:40 by aoo               #+#    #+#             */
-/*   Updated: 2025/03/21 03:06:41 by aoo              ###   ########.fr       */
+/*   Updated: 2025/04/04 03:41:44 by aoo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,6 +34,6 @@ class	Contact
 };
 
 void    get_safe_input(std::string prompt, std::string &value);
-void	print_setw(int width, std::string str);
+void	print_setw(size_t width, std::string str);
 
 #endif
